+++ algorithmtest/redblack_bst_tests.cpp
@@ -15,7 +15,7 @@ namespace algorithmtest
 
 		TEST_METHOD(RedBlackBSTTest1)
 		{
-			RedBlackBST<string, int> st;
+			RedBlackBST<string, size_t> st;
 			Assert::IsTrue(st.IsEmpty(), L"", LINE_INFO());
 
 			st.Put("asd",355);
@@ -37,7 +37,7 @@ namespace algorithmtest
 			Assert::IsFalse(st.Contains(key), L"", LINE_INFO());
 			Assert::IsFalse(st.Get(key).has_value(), L"", LINE_INFO());
 
-			int value = st.Get("is").value();
+			size_t value = st.Get("is").value();
 			Assert::IsTrue(value == 6, L"", LINE_INFO());
 			Assert::IsTrue(st.Select(3) == "quest", L"", LINE_INFO());
